requirements.txt
google-adk==1.5.0
mcp==1.9.1
deprecated==1.2.18